+++ src/styles.css
@@ -1,8 +1,8 @@
-body {
+/* body {
   width: 100%;
   padding: 0;
   margin: 0;
-}
+} */
 
 .tela-inteira {
   height: 100vh;
@@ -20,7 +20,7 @@ header {
 }
 
 main {
-  min-height: 80%;
+  /* min-height: 80%; */
   display: flex;
 }
 
@@ -38,7 +38,7 @@ main {
   flex-grow: 1;
   display: grid;
   grid-template-columns: repeat(4, 1fr);
-  grid-template-rows: repeat(4, 1fr);
+  grid-template-rows: repeat(2, 1fr);
   row-gap: 10px;
   column-gap: 10px;
   margin: 10px;
@@ -57,11 +57,11 @@ main {
 footer {
   background: #333b3e;
   color: white;
-  position: fixed;
-  bottom: 0;
+  /* position: fixed; */
+  /* bottom: 0; */
   width: 100%;
   display: flex;
-  padding: 0 10px;
+  /* padding: 10px 10px; */
 }
 
 .box-pagina-principal > img {
@@ -71,4 +71,15 @@ footer {
 
 h4 {
   margin-left: 20px;
+}
+
+.fotoUsuario {
+  width: 40px;
+  border-radius: 50px;
+}
+
+.infoUsuario {
+  margin: 4%;
+  display: flex;
+  align-items: center;
 }
